Refactored MovieTiles to MovieList for further use
