+++ src/App.js
@@ -2,7 +2,7 @@ import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
 import NavigationBar from "./common/NavigationBar";
 import Section from "./common/Section";
-import MoviesTiles from "./features/MoviesTiles";
+import MovieList from "./features/MovieList";
 import People from "./features/People";
 import Profile from "./features/Profile";
 import Pagination from "./common/Pagination";
@@ -34,7 +34,7 @@ function App() {
           <Route path="/movies">
             <Section
               title="Popular movies"
-              content={<MoviesTiles />}
+              content={<MovieList />}
               extraContent={<Pagination />}
             />
           </Route>
