+++ src/rootSaga.js
@@ -1,7 +1,7 @@
 import { all } from "redux-saga/effects";
 
 import { watchFetchMovieDetails } from "./features/MovieDetails/movieDetailsSaga";
-import { watchFetchMoviesList } from "./features/MoviesTiles/movieListSaga";
+import { watchFetchMoviesList } from "./features/MovieList/movieListSaga";
 import { watchFetchPeopleDetails } from "./features/Profile/peopleDetailsSaga";
 import { watchFetchPeopleList } from "./features/People/peopleSaga";
