+++ src/store.js
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./rootSaga";
 import movieDetailsReducer from "./features/MovieDetails/movieDetailsSlice";
-import moviesListReducer from "./features/MoviesTiles/movieListSlice";
+import moviesListReducer from "./features/MovieList/movieListSlice";
 import peopleListReducer from "./features/People/peopleSlice";
 import peopleDetailsReducer from "./features/Profile/peopleDetailsSlice";
 const sagaMiddleware = createSagaMiddleware();
